refactor: 컨벤션에 맞게 scss 파일명 수정
