+++ pages/sign-in/profile-setup/ProfileSetup.module.scss
